+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log (vscode-deploy-reloaded)
 
+## 0.7.0 (December 29th, 2017; npm)
+
+* bugfixes
+* code improvements
+* added `initNodeModules` [setting](https://github.com/mkloubert/vscode-deploy-reloaded/wiki#settings--), which runs `npm install` inside the workspace folder on startup, if a `package.json` file exists and NO `node_modules` sub folder has been found
+* added [tools](https://github.com/mkloubert/vscode-deploy-reloaded#npm-helpers-) for Node Package Manager ([npm](https://www.npmjs.com/package/npm))
+* added global and context based event properties (`events` and `globalEvents`) to [ScriptArguments](https://mkloubert.github.io/vscode-deploy-reloaded/interfaces/_contracts_.scriptarguments.html)
+
 ## 0.6.0 (December 28th, 2017; bugfixes)
 
 * fixed [folder mappings](https://github.com/mkloubert/vscode-deploy-reloaded/wiki/folder_mappings)